test(element): remove old tests & add new test for element decorator
