+++ packages/element/__tests__/index.ts
@@ -1,17 +1,5 @@
-import testCreateComputingEntanglement from './computed';
-import testCorpusculeElement from './corpusculeElement';
-import testDecorators from './decorators';
-import testDhtml from './dhtml';
-import testScheduler from './scheduler';
+import testElementDecorator from './decorators/element';
 
 describe('@corpuscule/element', () => {
-  afterEach(() => {
-    document.body.innerHTML = ''; // tslint:disable-line:no-inner-html
-  });
-
-  testDhtml();
-  testCorpusculeElement();
-  testDecorators();
-  testCreateComputingEntanglement();
-  testScheduler();
+  testElementDecorator();
 });
